+++ akamai-edgeworkers
@@ -2,4 +2,4 @@
 
 // stub file to kick off compiled app
 // top level CLI needs the file to be started with akamai
-require('./bin/src/edgeworkers/ewcommand.js');
+require('./bin/src/edgeworkers/ew-cli-main.js');
